+++ packages/frontend/src/app/(top-nav)/data-availability/projects/[layer]/_utils/get-da-throughput-section.ts
@@ -1,6 +1,6 @@
 import type { Project } from '@l2beat/config'
 import { featureFlags } from '~/consts/feature-flags'
-import { getDaThroughputTable } from '~/server/features/data-availability/utils/get-da-throuput'
+import { getDaThroughputTable } from '~/server/features/data-availability/utils/get-da-throuput-table'
 import { ps } from '~/server/projects'
 import { api } from '~/trpc/server'
 import { formatBytes } from '~/utils/number-format/format-bytes'
