+++ packages/frontend/src/server/features/data-availability/throughput/get-da-throughput-entries.ts
@@ -6,7 +6,7 @@ import type { CommonProjectEntry } from '../../utils/get-common-project-entry'
 import {
   type ThroughputTableData,
   getDaThroughputTable,
-} from '../utils/get-da-throuput'
+} from '../utils/get-da-throuput-table'
 import { getThroughputSyncWarning } from './is-throughput-synced'
 
 export async function getDaThroughputEntries(): Promise<DaThroughputEntry[]> {
